+++ CHANGELOG.md
@@ -8,6 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.10.0] - 2025-11-11
 
 ### Added
+- [Common] Add HTTP transport support (StreamableHTTP & legacy SSE) [#196](https://github.com/SmartBear/smartbear-mcp/pull/196)
+- [Common] Add centralized client registry system [#196](https://github.com/SmartBear/smartbear-mcp/pull/196)
+- [Common] Add common cache service [#196](https://github.com/SmartBear/smartbear-mcp/pull/196)
+
+### Changed
+- [Common] Refactor client authentication [#196](https://github.com/SmartBear/smartbear-mcp/pull/196)
 - [Qmetry] New QMetry MCP Server Tools Added, Refactored the existing tools structure [#217](https://github.com/SmartBear/smartbear-mcp/pull/217)
 - [Collaborator] Initial Collaborator client implementation [#223](https://github.com/SmartBear/smartbear-mcp/pull/223)